requirements.txt
streamlit
tensorflow==2.19.0
keras==3.0.5
pillow
numpy
